Introduced SocketReaderInitializer
