+++ hazelcast/src/main/java/com/hazelcast/nio/tcp/nonblocking/NonBlockingSocketReader.java
@@ -16,28 +16,23 @@
 
 package com.hazelcast.nio.tcp.nonblocking;
 
+import com.hazelcast.internal.metrics.Probe;
+import com.hazelcast.internal.util.counters.SwCounter;
 import com.hazelcast.logging.ILogger;
-import com.hazelcast.nio.IOService;
-import com.hazelcast.nio.Protocols;
-import com.hazelcast.nio.ascii.TextReadHandler;
-import com.hazelcast.nio.tcp.ClientReadHandler;
+import com.hazelcast.nio.Connection;
 import com.hazelcast.nio.tcp.ReadHandler;
+import com.hazelcast.nio.tcp.SocketChannelWrapper;
 import com.hazelcast.nio.tcp.SocketReader;
-import com.hazelcast.nio.tcp.SocketWriter;
-import com.hazelcast.nio.tcp.TcpIpConnection;
-import com.hazelcast.nio.tcp.TcpIpConnectionManager;
+import com.hazelcast.nio.tcp.SocketReaderInitializer;
 import com.hazelcast.nio.tcp.nonblocking.iobalancer.IOBalancer;
 
 import java.io.EOFException;
 import java.io.IOException;
-import java.net.SocketException;
 import java.nio.ByteBuffer;
 
-import static com.hazelcast.nio.ConnectionType.MEMBER;
-import static com.hazelcast.nio.IOService.KILO_BYTE;
-import static com.hazelcast.nio.Protocols.CLIENT_BINARY_NEW;
-import static com.hazelcast.nio.Protocols.CLUSTER;
-import static com.hazelcast.util.StringUtil.bytesToString;
+import static com.hazelcast.internal.util.counters.SwCounter.newSwCounter;
+import static java.lang.System.currentTimeMillis;
+import static java.nio.channels.SelectionKey.OP_READ;
 
 /**
  * A {@link SocketReader} tailored for non blocking IO.
@@ -46,71 +41,179 @@ import static com.hazelcast.util.StringUtil.bytesToString;
  * {@link #handle()} is called to read out the data from the socket into a bytebuffer and hand it over to the
  * {@link ReadHandler} to get processed.
  */
-public final class NonBlockingSocketReader
-        extends AbstractNonBlockingSocketReader<TcpIpConnection> {
+public final class NonBlockingSocketReader<C extends Connection>
+        extends AbstractHandler<C>
+        implements SocketReader {
 
-    private final TcpIpConnectionManager connectionManager;
-    private final IOService ioService;
+    protected ByteBuffer inputBuffer;
+
+    @Probe(name = "bytesRead")
+    private final SwCounter bytesRead = newSwCounter();
+    @Probe(name = "normalFramesRead")
+    private final SwCounter normalFramesRead = newSwCounter();
+    @Probe(name = "priorityFramesRead")
+    private final SwCounter priorityFramesRead = newSwCounter();
+
+    private ReadHandler readHandler;
+    private volatile long lastReadTime;
+    private final SocketReaderInitializer initializer;
+    private final ByteBuffer protocolBuffer = ByteBuffer.allocate(3);
 
     public NonBlockingSocketReader(
-            TcpIpConnection connection,
+            C connection,
+            SocketChannelWrapper socketChannel,
             NonBlockingIOThread ioThread,
             ILogger logger,
-            IOBalancer balancer) {
-        super(connection, connection.getSocketChannelWrapper(), ioThread, logger, balancer);
-
-        this.connectionManager = connection.getConnectionManager();
-        this.ioService = connectionManager.getIoService();
+            IOBalancer balancer,
+            SocketReaderInitializer initializer) {
+        super(connection, ioThread, OP_READ, socketChannel, logger, balancer);
+        this.initializer = initializer;
     }
 
     @Override
-    protected ReadHandler initReadHandler() throws IOException {
-        ByteBuffer protocolBuffer = ByteBuffer.allocate(3);
-        int readBytes = socketChannel.read(protocolBuffer);
-        if (readBytes == -1) {
-            throw new EOFException("Could not read protocol type!");
-        }
+    public ByteBuffer getProtocolBuffer() {
+        return protocolBuffer;
+    }
 
-        if (readBytes == 0 && connectionManager.isSSLEnabled()) {
-            // when using SSL, we can read 0 bytes since data read from socket can be handshake frames.
-            return null;
-        }
+    @Override
+    public void initInputBuffer(ByteBuffer inputBuffer) {
+        this.inputBuffer = inputBuffer;
+    }
 
-        ReadHandler readHandler = null;
-        if (!protocolBuffer.hasRemaining()) {
-            String protocol = bytesToString(protocolBuffer.array());
-            SocketWriter socketWriter = connection.getSocketWriter();
-            if (CLUSTER.equals(protocol)) {
-                configureBuffers(ioService.getSocketReceiveBufferSize() * KILO_BYTE);
-                connection.setType(MEMBER);
-                socketWriter.setProtocol(CLUSTER);
-                readHandler = ioService.createReadHandler(connection);
-            } else if (CLIENT_BINARY_NEW.equals(protocol)) {
-                configureBuffers(ioService.getSocketClientReceiveBufferSize() * KILO_BYTE);
-                socketWriter.setProtocol(CLIENT_BINARY_NEW);
-                readHandler = new ClientReadHandler(getNormalFramesReadCounter(), connection, ioService);
-            } else {
-                configureBuffers(ioService.getSocketReceiveBufferSize() * KILO_BYTE);
-                socketWriter.setProtocol(Protocols.TEXT);
-                inputBuffer.put(protocolBuffer.array());
-                readHandler = new TextReadHandler(connection);
-                connection.getConnectionManager().incrementTextConnections();
+    @Override
+    public void initReadHandler(ReadHandler readHandler) {
+        this.readHandler = readHandler;
+    }
+
+    @Probe(name = "idleTimeMs")
+    private long idleTimeMs() {
+        return Math.max(currentTimeMillis() - lastReadTime, 0);
+    }
+
+    @Override
+    public SwCounter getNormalFramesReadCounter() {
+        return normalFramesRead;
+    }
+
+    @Override
+    public SwCounter getPriorityFramesReadCounter() {
+        return priorityFramesRead;
+    }
+
+    @Override
+    public long getLastReadTimeMillis() {
+        return lastReadTime;
+    }
+
+    @Override
+    public void init() {
+        ioThread.addTaskAndWakeup(new Runnable() {
+            @Override
+            public void run() {
+                try {
+                    getSelectionKey();
+                } catch (Throwable t) {
+                    onFailure(t);
+                }
+            }
+        });
+    }
+
+    /**
+     * Migrates this handler to a new NonBlockingIOThread.
+     * The migration logic is rather simple:
+     * <p><ul>
+     * <li>Submit a de-registration task to a current NonBlockingIOThread</li>
+     * <li>The de-registration task submits a registration task to the new NonBlockingIOThread</li>
+     * </ul></p>
+     *
+     * @param newOwner target NonBlockingIOThread this handler migrates to
+     */
+    @Override
+    public void requestMigration(NonBlockingIOThread newOwner) {
+        ioThread.addTaskAndWakeup(new StartMigrationTask(newOwner));
+    }
+
+    @Override
+    public void handle() throws Exception {
+        eventCount.inc();
+        // we are going to set the timestamp even if the socketChannel is going to fail reading. In that case
+        // the connection is going to be closed anyway.
+        lastReadTime = currentTimeMillis();
+
+        if (readHandler == null) {
+            initializer.init(connection, this);
+            if (readHandler == null) {
+                // when using SSL, we can read 0 bytes since data read from socket can be handshake frames.
+                return;
             }
         }
 
-        if (readHandler == null) {
-            throw new IOException("Could not initialize ReadHandler!");
+        int readBytes = socketChannel.read(inputBuffer);
+        if (readBytes <= 0) {
+            if (readBytes == -1) {
+                throw new EOFException("Remote socket closed!");
+            }
+            return;
         }
-        return readHandler;
+
+        bytesRead.inc(readBytes);
+
+        inputBuffer.flip();
+        readHandler.onRead(inputBuffer);
+        if (inputBuffer.hasRemaining()) {
+            inputBuffer.compact();
+        } else {
+            inputBuffer.clear();
+        }
     }
 
-    private void configureBuffers(int size) {
-        super.configureInputBuffer(size, ioService.isSocketBufferDirect());
+    @Override
+    public void close() {
+        ioThread.addTaskAndWakeup(new Runnable() {
+            @Override
+            public void run() {
+                if (ioThread != Thread.currentThread()) {
+                    // the NonBlockingSocketReader has migrated to a different IOThread after the close got called.
+                    // so we need to send the task to the right ioThread. Otherwise multiple ioThreads could be accessing
+                    // the same socketChannel.
+                    ioThread.addTaskAndWakeup(this);
+                    return;
+                }
 
-        try {
-            connection.setReceiveBufferSize(size);
-        } catch (SocketException e) {
-            logger.finest("Failed to adjust TCP receive buffer of " + connection + " to " + size + " B.", e);
+                try {
+                    socketChannel.closeInbound();
+                } catch (IOException e) {
+                    logger.finest("Error while closing inbound", e);
+                }
+            }
+        });
+    }
+
+    @Override
+    public String toString() {
+        return connection + ".socketReader";
+    }
+
+    private class StartMigrationTask implements Runnable {
+        private final NonBlockingIOThread newOwner;
+
+        StartMigrationTask(NonBlockingIOThread newOwner) {
+            this.newOwner = newOwner;
+        }
+
+        @Override
+        public void run() {
+            // if there is no change, we are done
+            if (ioThread == newOwner) {
+                return;
+            }
+
+            try {
+                startMigration(newOwner);
+            } catch (Throwable t) {
+                onFailure(t);
+            }
         }
     }
 }
